+++ Util.h
@@ -11,10 +11,10 @@ BOOL WriteToReg_Int(HKEY hKey, LPCTSTR lpKey, LPCTSTR lpValue, UINT nData);
 // 문자열을 쓴다.
 BOOL WriteToReg_String(HKEY hKey, LPCTSTR lpKey, LPCTSTR lpValue, LPCTSTR lpData);
 
-void ReadFromIniFile_Int(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, int nDefault, int &nReturn);
+void ReadFromIniFile_Int(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, long nDefault, long &nReturn);
 void ReadFromIniFile_Float(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, float fDefault, float &fReturn);
 void ReadFromIniFile_String(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, const CString &strDefault, CString &strReturn);
-void WriteToIniFile_Int(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, int nValue);
+void WriteToIniFile_Int(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, long nValue);
 void WriteToIniFile_Float(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, float fValue);
 void WriteToIniFile_String(LPCTSTR szPathName, LPCTSTR szSection, LPCTSTR szKeyName, const CString &strValue);
 
@@ -34,12 +34,12 @@ void GetFilePathListIncludingSubFolder(CString strPathName, CString strFileName,
 void GetDirTitleList(CString aszPathName, CString aszFileName, OUT CArray<CString> &arDirTitles);
 void GetDirPathList(CString aszPathName, CString aszFileName, OUT CArray<CString> &arDirPaths);
 
-BOOL ExtractSubString(CString& rString, LPCTSTR lpszFullString, int iSubString, TCHAR chSep = '\n');
-BOOL ExtractSublong(long& lValue, LPCTSTR lpszFullString, int iSubString, TCHAR chSep = '\n');
+BOOL ExtractSubString(CString& rString, LPCTSTR lpszFullString, long iSubString, TCHAR chSep = '\n');
+BOOL ExtractSublong(long& lValue, LPCTSTR lpszFullString, long iSubString, TCHAR chSep = '\n');
 
 BOOL ResourceToFile(long lResourceID ,CString strResourceType , CString strFullPath);
 
-void DeleteOldFiles(CString strPathName, CString strFileName, int nLeaveDaysAgo, BOOL bDeleteSubFolder = FALSE, BOOL bDeleteEmptyFolder = FALSE, CTime tmToday = CTime(CTime::GetCurrentTime().GetYear(), CTime::GetCurrentTime().GetMonth(), CTime::GetCurrentTime().GetDay(),0,0,0) );
+void DeleteOldFiles(CString strPathName, CString strFileName, long nLeaveDaysAgo, BOOL bDeleteSubFolder = FALSE, BOOL bDeleteEmptyFolder = FALSE, CTime tmToday = CTime(CTime::GetCurrentTime().GetYear(), CTime::GetCurrentTime().GetMonth(), CTime::GetCurrentTime().GetDay(),0,0,0) );
 
 long GetCpuUsagePercent();
